[Feat] Find Actual Ratio

+++ app/src/main/java/com/example/demoidimgktx/model/FrameInfo.kt
@@ -1,16 +1,16 @@
 package com.example.demoidimgktx.model
 
 data class FrameInfo(
-    val x: Int,
-    val y: Int,
+    val x: Float,
+    val y: Float,
     val rotation : Float = 0F,
-    val width: Int,
-    val height: Int
+    val width: Float,
+    val height: Float
 )
 
 data class FramesMeta(
-    val width: Int,
-    val height: Int,
+    val width: Float,
+    val height: Float,
     val foregroundPath : String,
     val frameList: List<FrameInfo>
 )
